Move Polit.* and Faction.* in Galaxy; Remove test* ; Reword includes in involved files; Some decouple
