+++ src/LuaFaction.cpp
@@ -1,7 +1,7 @@
 // Copyright © 2008-2019 Pioneer Developers. See AUTHORS.txt for details
 // Licensed under the terms of the GPL v3. See licenses/GPL-3.txt
 
-#include "Factions.h"
+#include "galaxy/Factions.h"
 #include "LuaConstants.h"
 #include "LuaObject.h"
 #include "LuaUtils.h"
