+++ src/LuaStarSystem.cpp
@@ -2,7 +2,6 @@
 // Licensed under the terms of the GPL v3. See licenses/GPL-3.txt
 
 #include "EnumStrings.h"
-#include "Factions.h"
 #include "FileSystem.h"
 #include "Game.h"
 #include "GameSaveError.h"
@@ -16,6 +15,7 @@
 #include "SpaceStation.h"
 #include "Star.h"
 #include "galaxy/Economy.h"
+#include "galaxy/Factions.h"
 #include "galaxy/Galaxy.h"
 #include "galaxy/GalaxyCache.h"
 #include "galaxy/Sector.h"
