+++ src/enum_table.cpp
@@ -9,13 +9,13 @@
 #include "LuaEngine.h"
 #include "LuaFileSystem.h"
 #include "Object.h"
-#include "Polit.h"
 #include "Propulsion.h"
 #include "Ship.h"
 #include "ShipAICmd.h"
 #include "ShipType.h"
 #include "galaxy/Economy.h"
 #include "galaxy/StarSystem.h"
+#include "galaxy/Polit.h"
 #include "gameui/Face.h"
 #include "gameui/LabelOverlay.h"
 #include "scenegraph/Model.h"
