+++ src/galaxy/Factions.cpp
