+++ src/galaxy/Factions.h
