+++ src/galaxy/Polit.cpp
